7/11 日常确认合并

+++ book_builder.py
@@ -1,14 +1,19 @@
 import os
 
 
-def build_all_dir(dir_path, file_name):
+def build_project(dir_path: str, project_name: str) -> None:
     os.chdir(dir_path)
     os.system(
-        "xelatex --no-pdf -synctex=1 -interaction=nonstopmode -file-line-erro -shell-escape " + file_name)
+        "xelatex --no-pdf -synctex=1 -interaction=nonstopmode -file-line-erro -shell-escape " + project_name + ".tex")
     os.system(
-        "xelatex -synctex=1 -interaction=nonstopmode -file-line-erro -shell-escape " + file_name)
-    os.system("rm *.bcf *.out *.xml *.toc *.xdv")
+        "xelatex --no-pdf -synctex=1 -interaction=nonstopmode -file-line-erro -shell-escape " + project_name + ".tex")
+    os.system(
+        "xdvipdfmx -E " + project_name + ".xdv")
+    if os.name == "nt":
+        os.system("del *.bcf *.out *.xml *.toc *.xdv")
+    else:
+        os.system("rm *.bcf *.out *.xml *.toc *.xdv")
 
 
 if __name__ == "__main__":
-    build_all_dir("书籍/基础篇/", "基础篇.tex")
+    build_project("书籍/基础篇/", "基础篇")
